Changes: 1. Added bow attack 2. Expanded ground
Notes:
- For testing purposes, the ground has been expanded
- To shoot the bow, press the F key
