--- Assets/Attack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f36731c1aab11ba4393dc662977215da
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
