+++ Assets/Characters/Player/Player.prefab
@@ -210,6 +210,37 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   attackDamage: 15
   knockback: {x: 1, y: 1}
+--- !u!1 &4424320375762098309
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 5201776749741681013}
+  m_Layer: 7
+  m_Name: BowAttack
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &5201776749741681013
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4424320375762098309}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0.51, y: -0.19, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 6275902163922745359}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &4596089932920384812
 GameObject:
   m_ObjectHideFlags: 0
@@ -227,6 +258,7 @@ GameObject:
   - component: {fileID: 5509283050027886590}
   - component: {fileID: 7073416263497750709}
   - component: {fileID: 3149450074100316420}
+  - component: {fileID: 3068370490356732770}
   m_Layer: 7
   m_Name: Player
   m_TagString: Untagged
@@ -251,6 +283,7 @@ Transform:
   - {fileID: 5495437338773776327}
   - {fileID: 8703813309057519228}
   - {fileID: 6659827120149866315}
+  - {fileID: 5201776749741681013}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &3364222829578848417
@@ -457,6 +490,22 @@ MonoBehaviour:
         m_CallState: 2
     m_ActionId: 6159fad5-3825-4d85-bf51-3b57e9213d32
     m_ActionName: Player/Jump[/Keyboard/space]
+  - m_PersistentCalls:
+      m_Calls:
+      - m_Target: {fileID: 992949383336453805}
+        m_TargetAssemblyTypeName: PlayerController, Assembly-CSharp
+        m_MethodName: OnRangedAttack
+        m_Mode: 0
+        m_Arguments:
+          m_ObjectArgument: {fileID: 0}
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
+          m_IntArgument: 0
+          m_FloatArgument: 0
+          m_StringArgument: 
+          m_BoolArgument: 0
+        m_CallState: 2
+    m_ActionId: d07b46ed-8df4-47ff-b9c8-42553593329d
+    m_ActionName: Player/RangedAttack[/Keyboard/f]
   m_NeverAutoSwitchControlSchemes: 0
   m_DefaultControlScheme: 
   m_DefaultActionMap: Player
@@ -609,6 +658,20 @@ MonoBehaviour:
   isAlive: 1
   isInvincible: 0
   invincibilityTimer: 1
+--- !u!114 &3068370490356732770
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4596089932920384812}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8e4355f38fa1fda41a3c44d27af1c6c5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  launchPoint: {fileID: 5201776749741681013}
+  projectilePrefab: {fileID: 2441554554683442036, guid: d88fb9b8f21f2004585390f568c94607, type: 3}
 --- !u!1 &6529107816059400107
 GameObject:
   m_ObjectHideFlags: 0
