+++ Assets/Characters/Player/player_bow.anim
@@ -1,0 +1,103 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!74 &7400000
+AnimationClip:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: player_bow
+  serializedVersion: 7
+  m_Legacy: 0
+  m_Compressed: 0
+  m_UseHighQualityCurve: 1
+  m_RotationCurves: []
+  m_CompressedRotationCurves: []
+  m_EulerCurves: []
+  m_PositionCurves: []
+  m_ScaleCurves: []
+  m_FloatCurves: []
+  m_PPtrCurves:
+  - serializedVersion: 2
+    curve:
+    - time: 0
+      value: {fileID: 21300000, guid: 8aca4893c9fd1174a87d9716979a21c3, type: 3}
+    - time: 0.16666667
+      value: {fileID: 21300000, guid: e57a1869114af5c45918bfd845429df3, type: 3}
+    - time: 0.33333334
+      value: {fileID: 21300000, guid: d382a81ff32b496448e216320cedca3f, type: 3}
+    - time: 0.5
+      value: {fileID: 21300000, guid: fa4826329a751d94484a760a480f851e, type: 3}
+    - time: 0.6666667
+      value: {fileID: 21300000, guid: 20ad97eeeb4a8c34b99c092c888ed73d, type: 3}
+    - time: 0.8333333
+      value: {fileID: 21300000, guid: b442aa523dab38447ac597d6caf5ab93, type: 3}
+    - time: 1
+      value: {fileID: 21300000, guid: b3c04927acc37fe47974fc3f6538781e, type: 3}
+    - time: 1.1666666
+      value: {fileID: 21300000, guid: d58bd4f8a2e55a24292990e96841aa2d, type: 3}
+    - time: 1.3333334
+      value: {fileID: 21300000, guid: 9545d71eb551b514687876d35c18b444, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+    flags: 2
+  m_SampleRate: 60
+  m_WrapMode: 0
+  m_Bounds:
+    m_Center: {x: 0, y: 0, z: 0}
+    m_Extent: {x: 0, y: 0, z: 0}
+  m_ClipBindingConstant:
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+      isIntCurve: 0
+      isSerializeReferenceCurve: 0
+    pptrCurveMapping:
+    - {fileID: 21300000, guid: 8aca4893c9fd1174a87d9716979a21c3, type: 3}
+    - {fileID: 21300000, guid: e57a1869114af5c45918bfd845429df3, type: 3}
+    - {fileID: 21300000, guid: d382a81ff32b496448e216320cedca3f, type: 3}
+    - {fileID: 21300000, guid: fa4826329a751d94484a760a480f851e, type: 3}
+    - {fileID: 21300000, guid: 20ad97eeeb4a8c34b99c092c888ed73d, type: 3}
+    - {fileID: 21300000, guid: b442aa523dab38447ac597d6caf5ab93, type: 3}
+    - {fileID: 21300000, guid: b3c04927acc37fe47974fc3f6538781e, type: 3}
+    - {fileID: 21300000, guid: d58bd4f8a2e55a24292990e96841aa2d, type: 3}
+    - {fileID: 21300000, guid: 9545d71eb551b514687876d35c18b444, type: 3}
+  m_AnimationClipSettings:
+    serializedVersion: 2
+    m_AdditiveReferencePoseClip: {fileID: 0}
+    m_AdditiveReferencePoseTime: 0
+    m_StartTime: 0
+    m_StopTime: 1.35
+    m_OrientationOffsetY: 0
+    m_Level: 0
+    m_CycleOffset: 0
+    m_HasAdditiveReferencePose: 0
+    m_LoopTime: 0
+    m_LoopBlend: 0
+    m_LoopBlendOrientation: 0
+    m_LoopBlendPositionY: 0
+    m_LoopBlendPositionXZ: 0
+    m_KeepOriginalOrientation: 0
+    m_KeepOriginalPositionY: 1
+    m_KeepOriginalPositionXZ: 0
+    m_HeightFromFeet: 0
+    m_Mirror: 0
+  m_EditorCurves: []
+  m_EulerEditorCurves: []
+  m_HasGenericRootTransform: 0
+  m_HasMotionFloatCurves: 0
+  m_Events:
+  - time: 1.1666666
+    functionName: FireProjectile
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
