+++ Assets/Characters/Player/player_bow.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63eaf8405a374064f8bd3d244d6e865d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
