--- Assets/FadeRemoveBehaviour.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7444fac165119cb4d837d4aac42a61d4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
