+++ Assets/Objects/Projectiles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 54ea87e338cf50e43ab4d2dcb5d174a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
