+++ Assets/Scripts/AnimationStrings.cs
@@ -23,5 +23,6 @@ internal class AnimationStrings
     internal static string hitTrigger = "Hit";
     internal static string lockVelocity = "LockVelocity";
     internal static string attackCooldown = "AttackCooldown";
+    internal static string rangedAttackTrigger = "RangedAttack";
 }
