+++ Assets/Scripts/Damageable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98546adaff3c0d549bd9c2db483f02c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
