+++ Assets/Scripts/ProjectileLauncher.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e4355f38fa1fda41a3c44d27af1c6c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
