--- Assets/SetFloatBehaviour.cs
@@ -1,68 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class SetFloatBehaviour : StateMachineBehaviour
-{
-    public string floatName;
-    public bool updateOnStateEnter, updateOnStateExit;
-    public bool updateOnStateMachineEnter, updateOnStateMachineExit;
-    public float valueOnEnter, valueOnExit;
-
-    // OnStateEnter is called before OnStateEnter is called on any state inside this state machine
-    override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    {
-        if (updateOnStateEnter)
-        {
-            animator.SetFloat(floatName, valueOnEnter);
-        }
-    }
-
-    // OnStateUpdate is called before OnStateUpdate is called on any state inside this state machine
-    //override public void OnStateUpdate(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    //{
-    //    // TODO: Decide whether this is still needed
-    //    // Check if the attack button is pressed again
-    //    if (Input.GetMouseButton(0))
-    //    {
-    //        Debug.Log("Left mouse button clicked! (OnStateUpdate)");
-    //        // Reset and set the attack trigger to cancel and restart the animation
-    //        animator.ResetTrigger(AnimationStrings.attackTrigger);
-    //    }
-    //}
-
-    // OnStateExit is called before OnStateExit is called on any state inside this state machine
-    override public void OnStateExit(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    {
-        if (updateOnStateExit)
-        {
-            animator.SetFloat(floatName, valueOnExit);
-        }
-    }
-
-    // OnStateMove is called before OnStateMove is called on any state inside this state machine
-    //override public void OnStateMove(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    //{
-    //    
-    //}
-
-    // OnStateIK is called before OnStateIK is called on any state inside this state machine
-    //override public void OnStateIK(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    //{
-    //    
-    //}
-
-    // OnStateMachineEnter is called when entering a state machine via its Entry Node
-    override public void OnStateMachineEnter(Animator animator, int stateMachinePathHash)
-    {
-        if (updateOnStateMachineEnter)
-            animator.SetFloat(floatName, valueOnEnter);
-    }
-
-    // OnStateMachineExit is called when exiting a state machine via its Exit Node
-    override public void OnStateMachineExit(Animator animator, int stateMachinePathHash)
-    {
-        if (updateOnStateMachineExit)
-            animator.SetFloat(floatName, valueOnExit);
-    }
-}
