+++ Builds/PC/Prototype Platformer_Data/boot.config
@@ -3,4 +3,4 @@ gfx-enable-native-gfx-jobs=1
 wait-for-native-debugger=0
 hdr-display-enabled=0
 gc-max-time-slice=3
-build-guid=70f20b6c7eed4780b9cdd035878ac7de
+build-guid=a8df581f5c2b44f085a6bd5de4a6f8d8
