+++ ProjectSettings/ProjectSettings.asset
@@ -139,7 +139,7 @@ PlayerSettings:
   loadStoreDebugModeEnabled: 0
   visionOSBundleVersion: 1.0
   tvOSBundleVersion: 1.0
-  bundleVersion: 1.14.0-alpha
+  bundleVersion: 1.15.0-alpha
   preloadedAssets: []
   metroInputSource: 0
   wsaTransparentSwapchain: 0
